BJ-935: refactor properties
add different endpoints for card ranges and transaction info
replace preparation on platform
