+++ src/main/java/com/rbkmoney/threeds/server/config/rbkmoneyplatform/ResponseHandlerConfig.java
@@ -14,7 +14,7 @@ import org.springframework.context.annotation.Configuration;
 import static com.rbkmoney.threeds.server.config.builder.HandlerBuilder.createResponseHandler;
 
 @Configuration
-@ConditionalOnProperty(name = "preparation-flow.mode", havingValue = "RBK_MONEY_PLATFORM")
+@ConditionalOnProperty(name = "platform.mode", havingValue = "RBK_MONEY_PLATFORM")
 public class ResponseHandlerConfig {
 
     @Bean
