+++ src/main/java/com/rbkmoney/threeds/server/constants/PlatformMode.java
@@ -1,6 +1,6 @@
 package com.rbkmoney.threeds.server.constants;
 
-public enum PreparationFlowMode {
+public enum PlatformMode {
 
     RBK_MONEY_PLATFORM,
     TEST_PLATFORM
