+++ src/main/java/com/rbkmoney/threeds/server/handle/constraint/parq/PArqNotRequiredEnumGarbageWrapperContentConstraintValidationHandlerImpl.java
@@ -6,7 +6,6 @@ import com.rbkmoney.threeds.server.domain.root.proprietary.PArq;
 import com.rbkmoney.threeds.server.domain.threedsrequestor.ThreeDSRequestorAuthenticationInd;
 import com.rbkmoney.threeds.server.domain.threedsrequestor.ThreeRIInd;
 import com.rbkmoney.threeds.server.dto.ConstraintValidationResult;
-import com.rbkmoney.threeds.server.utils.Wrappers;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -14,6 +13,7 @@ import static com.rbkmoney.threeds.server.domain.threedsrequestor.ThreeDSRequest
 import static com.rbkmoney.threeds.server.domain.threedsrequestor.ThreeDSRequestorAuthenticationInd.RECURRING_TRANSACTION;
 import static com.rbkmoney.threeds.server.dto.ConstraintType.PATTERN;
 import static com.rbkmoney.threeds.server.utils.Wrappers.getGarbageValue;
+import static com.rbkmoney.threeds.server.utils.Wrappers.getValue;
 
 @Component
 @RequiredArgsConstructor
@@ -26,8 +26,8 @@ public class PArqNotRequiredEnumGarbageWrapperContentConstraintValidationHandler
 
     @Override
     public ConstraintValidationResult handle(PArq o) {
-        DeviceChannel deviceChannel = Wrappers.getValue(o.getDeviceChannel());
-        MessageCategory messageCategory = Wrappers.getValue(o.getMessageCategory());
+        DeviceChannel deviceChannel = getValue(o.getDeviceChannel());
+        MessageCategory messageCategory = getValue(o.getMessageCategory());
 
         if (getGarbageValue(o.getAddrMatch()) != null) {
             return ConstraintValidationResult.failure(PATTERN, "addrMatch");
@@ -38,24 +38,24 @@ public class PArqNotRequiredEnumGarbageWrapperContentConstraintValidationHandler
             return ConstraintValidationResult.failure(PATTERN, "browserColorDepth");
         }
 
-        Object purchaseDateGarbageValue = Wrappers.getGarbageValue(o.getPurchaseDate());
-        if (!(Wrappers.getValue(o.getThreeRIInd()) == ThreeRIInd.INSTALMENT_TRANSACTION)) {
+        Object purchaseDateGarbageValue = getGarbageValue(o.getPurchaseDate());
+        if (!(getValue(o.getThreeRIInd()) == ThreeRIInd.INSTALMENT_TRANSACTION)) {
             if (messageCategory != MessageCategory.NON_PAYMENT_AUTH && purchaseDateGarbageValue != null) {
                 return ConstraintValidationResult.failure(PATTERN, "purchaseDate");
             }
 
-            if (Wrappers.getGarbageValue(o.getRecurringExpiry()) != null) {
+            if (getGarbageValue(o.getRecurringExpiry()) != null) {
                 return ConstraintValidationResult.failure(PATTERN, "recurringExpiry");
             }
         }
 
-        ThreeDSRequestorAuthenticationInd authenticationInd = Wrappers.getValue(o.getThreeDSRequestorAuthenticationInd());
+        ThreeDSRequestorAuthenticationInd authenticationInd = getValue(o.getThreeDSRequestorAuthenticationInd());
         if (!(authenticationInd == INSTALMENT_TRANSACTION || authenticationInd == RECURRING_TRANSACTION)) {
             if (messageCategory != MessageCategory.NON_PAYMENT_AUTH && purchaseDateGarbageValue != null) {
                 return ConstraintValidationResult.failure(PATTERN, "purchaseDate");
             }
 
-            if (Wrappers.getGarbageValue(o.getRecurringExpiry()) != null) {
+            if (getGarbageValue(o.getRecurringExpiry()) != null) {
                 return ConstraintValidationResult.failure(PATTERN, "recurringExpiry");
             }
         }
@@ -98,11 +98,11 @@ public class PArqNotRequiredEnumGarbageWrapperContentConstraintValidationHandler
             return ConstraintValidationResult.failure(PATTERN, "payTokenSource");
         }
 
-        if (Wrappers.getValue(o.getWhiteListStatusSource()) == null && getGarbageValue(o.getWhiteListStatus()) != null) {
+        if (getValue(o.getWhiteListStatusSource()) == null && getGarbageValue(o.getWhiteListStatus()) != null) {
             return ConstraintValidationResult.failure(PATTERN, "whiteListStatus");
         }
 
-        if (Wrappers.getValue(o.getWhiteListStatus()) == null && getGarbageValue(o.getWhiteListStatusSource()) != null) {
+        if (getValue(o.getWhiteListStatus()) == null && getGarbageValue(o.getWhiteListStatusSource()) != null) {
             return ConstraintValidationResult.failure(PATTERN, "whiteListStatusSource");
         }
