+++ src/test/java/com/rbkmoney/threeds/server/flow/RBKMoneyGetChallengeFlowTest.java
@@ -25,7 +25,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
         classes = {ThreeDsServerApplication.class, MockConfig.class},
         properties = {
                 "spring.main.allow-bean-definition-overriding=true",
-                "preparation-flow.mode=RBK_MONEY_PLATFORM"})
+                "platform.mode=RBK_MONEY_PLATFORM"})
 @AutoConfigureMockMvc
 public class RBKMoneyGetChallengeFlowTest extends TestBase {
